Fix error string table bounds checks in NDEBUG builds
We were relying on asserts for critical bounds checks; this change ensures that
they are never optimized out. It also avoids crashing if junk is passed as an
error code to the error code lookup functions; we'd rather return an "Unknown
error code" string than kill the process if an error code is munged to be
negative somewhere.