Fix lookup3 valgrind warnings
This switches lookup3 to use a marginally slower way of accessing the last 12
bytes of a string, in exchange for removing the valgrind
access-beyond-end-of-buffer warnings that arise from its normal methods.

While this is a slight pessimization, leaving this optimization in makes it
difficult for consuming applications and libraries to perform valgrind testing;
valgrind appears to attribute the uninitialized accesses to later code, making
it very difficult to accurately suppress these warnings. Considering cases
where the aws-c-common library builder may not be the same person as the
library user, it's better to have this in the valgrind-safe mode by default; we
can later make it togglable if the performance gains are needed in
ultra-high-performance systems.