task scheduler revamp (#128)
Scheduler revamped so that it can be relied on not to fail.

It was difficult to build systems that used the scheduler, when there was always the possibility that a task would fail to schedule. By storing aws_task in a linked-list most of the time, and forcing users to manage the memory of the aws_task, we can guarantee that task scheduling will succeed.

The clock was also removed from the scheduler because, in reality, the user of the scheduler needed to keep another copy of the clock around for determining when to schedule tasks. By making the scheduler ignorant of where time comes from, it's simpler, easier to test, and more functions can be made infallible.

The "task popping" function was removed in favor of a simpler "has tasks" query.