Emit all public inlined APIs as symbols
This change ensures that all 'static inline' methods that are intended to be
public APIs are emitted as actual functions that can be invoked via a FFI,
while still retaining inlinability in C or C++ applications.

Additionally, this sets a SOVERSION and removes the undefined ${LIBTYPE}
variable reference in CMakeLists.txt. A shared library can be built by setting
the common cmake variable BUILD_SHARED_LIBS to ON.