Implement a SIMD-optimized constant-time base64 en/decoder (#127)
* Implement a SIMD-optimized constant-time base64 en/decoder

This change implements a parallel (24-byte-at-a-time) base64 en/decoder using
AVX2 instructions. Based on preliminary benchmarking, and depending
on CPU model, this provides a 3-4x speedup for both encode and decode for large
buffers; small buffers (<64 bytes) take double the time to process on my 2015
macbook, but on c5 class instances speed is comparable to the existing
implementation on the low end.

As a nice bonus, the new implementation avoids leaking data through cache
side-channels, as it does not use lookup tables for decoding.