Make most array_list methods inlinable
Most of the array list methods have very simple logic and are likely to be used
in performance-critical segments (eg, as part of the implementation of a
priority queue). This change makes them inlinable, to reduce such overhead.
It also adds restrict annotations to the array list parameter, to further reduce
spills due to false aliasing.