Add support for building on consoles
When cross-compiling to some platforms, the 'sysconf' API is not
available.
Also, we shouldn't attempt to build the tests since we can't run them
when cross-compiling.
Finally, some platforms use the GNU statement-expressions to implemnent
network-host byte order translations. Those extensions will generate a
warning when using the flag -Wgnu, this patch mitigates that.