Fix line wrapping of base64 decoding table.
The comment above this table indicates that it was intended to have 16
table entries per line; but in 4090d102 the table was automatically
rewrapped and ended up with 19 table entries per line.

Revert this to the intended formatting, and instruct clang-format to
ignore the table so that it doesn't get broken again.

Fixes:	#229