Updating Clock.c WIDEN macro to properly prepend the wide char idenfiier
The WIDEN2 macro was currently stringifying the input, but not concatenating the wchar_t identifier of L to the string.
This causes an issue if the input parameter was already stringified.
For example if the input parameter is `"Kernel32"`, the string `L"Kernel32"` would be formed.

The current logic works fine if the input parameter has not been stringified yet such as `"Kernel32"`