Test null-ness using AWS_NOT_NULL
Pre- and post-conditions previously mostly (though not always) relied on
the implicit cast from a pointer to an integral type. This might be ok
for a cast to _Bool, but cases where an assert expands to a function
expecting an int the results could be ambiguous.

Avoid any ambiguity by introducing a new macro AWS_NOT_NULL that
explicitly compares the pointer to NULL.