Fix proof helper declaration typing (#547)
* Fix proof helper declaration typing

The declarations of bounded_calloc, assert_byte_cursor_equivalence, and
assert_byte_buf_equivalence didn't match their definitions. Fix them to
avoid spurious compiler warnings.

Also fix one instance of mixing up 0 and NULL, and the typing of an
integer constant.

* Avoid compiler warnings when calling assert_{bytes_match,all_bytes_are}

Use explicit type casts to silence spurious warnings about incompatible
pointer types.

* Fix qsort stub declaration

Make sure the compare function has the expected type, and fix the typo
in the parameter name.

* Make lookup an array of bytes

The lookup table is expected to be an array of bytes, not an array of
pointers.

* Avoid compiler warnings about losing const-ness

Functions returning a const pointer should not lose that in the
assignment.

* Cleanup makefiles to avoid spurious warnings

In Makefiles, quotes aren't stripped the way they are by the shell.
Don't list object files twice. And enable compile-time warnings with
GOTO-CC.