Add empty source file for .inl-using proofs
Some proofs in C-Common don't link against any source files at all,
because the code that they are proving is written in .inl files that are
included by stubs. This commit adds an empty C file as the
PROJECT_SOURCES for those proofs so that the linking process proceeds
correctly.