Improved Posix Large File Support (LFS)
ISSUE:
Fixes bugs on 32bit Raspberry Pi OS where calls to aws_input_stream_seek() were failing. This was caused because the aws-c libraries were not properly enabling LFS, but the CPython was. The result was a disagreement over the size of `off_t`, which led to disagreement over the size of `aws_off_t` whose size was determined via ifdefs in a header.

SOLUTION:
Get conditional logic out of .h files and ensure all .c files are using LFS if possible.

1) Remove ~~aws_off_t~~, `int64_t` shouldl be used instead. This is a breaking API change. I will fix up downstream libraries in followup PRs.
2) AwsCFlags attempts to enable LFS, even on 32 bit system, by setting `-D_FILE_OFFSET_BITS=64` if necessary.