Fix Android Large File Support checks
32bit Android can have 64bit off_t, but no fseeko().

The check_c_source_compiles() sample program was supposed to check for this, but the missing fseeko() was only treated as a compiler warning and didn't fail the check. Therefore, explicitly check for fseeko as a separate check. (I couldn't figure out how to get check_c_source_compiles() to do warnings-as-errors)