Remove aws_ring_buffer_allocator_init declaration from public header
Make header not lie about the existence of `aws_ring_buffer_allocator_init` implementation.

I tried to call this API in my function but got an unresolved symbol.

 nm -gC crt/aws-c-common/libaws-c-common.a | grep aws_ring_buffer_
00000000000043be T aws_ring_buffer_check_atomic_ptr
00000000000043fe T aws_ring_buffer_is_empty
000000000000443f T aws_ring_buffer_is_valid
00000000000003d3 T aws_ring_buffer_acquire
000000000000095e T aws_ring_buffer_acquire_up_to
0000000000001363 T aws_ring_buffer_buf_belongs_to_pool
000000000000035c T aws_ring_buffer_clean_up
0000000000000216 T aws_ring_buffer_init
000000000000125b T aws_ring_buffer_release

And it's not there... looks like the implementation was recently removed but the declaration in the public header is still there.  This PR removes it.