Initial design of AWS JSON module
In theory, it should work as is, but there is a number of things that still need changing.
For example, the allocation of memory, additional tests, and the API still needs to be
actually used throughout the code base.

However, the majority of the code should, hopefully, be in place and good to go.