Hide symbols from libcrypto.a (#925)
**ISSUE:**
We were unable to take the latest version of AWS-LC (source of libcrypto.a).

AWS-LC changed the default visibility of symbols in static builds, from "hidden" to "visible" (see https://github.com/awslabs/aws-lc/pull/466)

This resulted in the symbols being visible in whatever shared library or executable ultimately used libcrypto.a. If an application ALSO ended up using libcrypto.so from the system's OpenSSL installation, we'd get crazy crashes due to SOME libcrypto function calls routing through libcrypto.a and SOME function calls routing through libcrypto.so.

**SOLUTION:**
Adapt the CMake function we use for configuring all C libraries in the Common Runtime ecosystem so that it always hides symbols from libcrypto.a.