Don't silently ignore sanitizers if requested but unavailable
Right now, if you set ENABLE_SANITIZERS but they're not available for
some reason, we keep chugging along as if nothing went wrong, except
that your build/tests won't actually have the sanitizers you asked for.
Because ENABLE_SANITIZERS is opt-in, we should instead fail loudly so
we don't silently run without the sanitizers you asked for.