Remove write race-condition in cJSON
Concurrent calls to `cJSON_Parse.*` routines mutually overwrite the `static` `global_error`
variable, which produces a write race condition caught by the thread sanitizer.

Since `cJSON_GetErrorPtr` is not used (as it is not thread-safe), it does not make sense
to fill in the value of the global variable and thus trigger the race condition.

A mutex would incur slow-down.

Hence remove write access to the `static` `global_error` variable from `cJSON_Parse.*` routines.

Resolves #953.