Add scripts/latest_submodules.py
Everyone has used this script for years to update submodules in the aws-crt-xyz repos. But we never had it in source control, because it wasn't clear where to put it. aws-c-common is a fine place. Now, from a CRT repo, you can run it like `./crt/aws-c-common/scripts/latest_submodules.py`

Also: Make the other scripts in aws-c-common/scripts executable.