Remove unused system: aws_bus (#996)
This system is unused. It was developed 2 years ago to support a [tracing system](https://github.com/awslabs/aws-c-common/tree/trace_event) that never got merged. It's a solid idea. We can always bring it back. But the tests are a bit flaky right now, and red code is best code.