Fix flaky test.
This test had two threads fighting to do the same repeated task, using a mutex to coordinate. At the end of the test, it asserted that each thread had done a bit of the work.

But rarely, 1 thread would do all the work before the other could even start. There was already code to make thread 1 wait for thread 2, but no code for thread 2 to wait for thread 1.

This should fix it.